requirements.txt
jupyter
pillow
torch
numpy
antspyx
torchvision
matplotlib
tqdm
scikit-learn
scikit-image
simpleitk
pandas
opencv-python
cuda-python
